Clean up some lingering inconsistent vertex/edge types

+++ fileConvert.cpp
@@ -37,7 +37,7 @@ int main(int argc, char *argv[]) {
   setUpFiles(argv[1], argv[2], fileIn, fileOut, inType, outType);
   bool keepReverseEdges = static_cast<bool>(atoi(argv[3]));
   bool isWeighted = false, isDirected = false, hasReverseEdges = false, isZeroIndexed = false, dropWeights = false;
-  int64_t numVerts = 0, numEdges = 0;
+  int32_t numVerts = 0, numEdges = 0;
   char *force_dw = std::getenv("CONVERT_FORCE_DROP_WEIGHTS");
   if (force_dw != nullptr) {
     std::cerr << "FORCE Drop Weights" << std::endl;
